update postcode regex to match addressing gateway validation. Move stray widget classes into appropriate module
