+++ application/presentation/widgets.py
@@ -4,6 +4,7 @@ from django.forms import widgets
 from django.utils.translation import gettext_lazy as _
 
 import govuk_forms.widgets as gov
+from govuk_forms.widgets import ChoiceWidget, RadioSelect
 
 
 class ExpirySplitDateWidget(gov.MultiWidget):
@@ -117,3 +118,13 @@ class SelectDateWidget(gov.MultiWidget):
         if value:
             return [value.month, value.year]
         return [None, None]
+
+
+class ConditionalPostChoiceWidget(ChoiceWidget):
+    template_name = 'multiple-select-post-conditional.html'
+    option_template_name = 'multiple-select-option-post-conditional.html'
+
+
+class ConditionalPostInlineRadioSelect(ConditionalPostChoiceWidget, RadioSelect):
+    field_group_classes = 'inline'
+    pass
