+++ nanny/settings/base.py
@@ -290,7 +290,7 @@ REGEX = {
              "5})|(?:\d{4}\)?[\s-]?(?:\d{5}|\d{3}[\s-]?\d{3}))|(?:\d{3}\)?[\s-]?\d{3}[\s-]?\d{3,4})|(?:\d{2}\)?["
              "\s-]?\d{4}[\s-]?\d{4}))(?:[\s-]?(?:x|ext\.?|\#)\d{3,4})?$",
     "INTERNATIONAL_PHONE": "^(\+|[0-9])[0-9]{5,20}$",
-    "POSTCODE_UPPERCASE": "^[A-Z]{1,2}[0-9]{1,2}[A-Z]?[0-9][A-Z][A-Z]$",
+    "POSTCODE_UPPERCASE": "^[A-Z]{1,2}[0-9][A-Z0-9]?[0-9][A-Z--CIKMOV]{2}$",
     "LAST_NAME": "^[A-zÀ-ÿ- ']+$",
     "MIDDLE_NAME": "^[A-zÀ-ÿ- ']+$",
     "FIRST_NAME": "^[A-zÀ-ÿ- ']+$",
